+++ Changelog.md
@@ -3,6 +3,7 @@
 ## V 1.5.0 - XX/XX/19
 
     - ControlComponent : Add MOUSEFOLLOW ControlType
+    - PhysicsComponent : Add MOUSEFOLLOW and MOUSECLICK CollisionCauses
     - Entry : Add Ctrl-C and Ctrl-V
     
     - Bug Fix : ControlComponent make weird movements with MOUSECLICK ControlType